feat: newsletters! or at least fetching and rendering
Squashed commit of the following:

commit c6d28398f7ff29160cef448c2bbc151080151fc1
Author: Adam Butler <ab@adam-butler.com>
Date:   Tue Apr 23 17:44:02 2024 +0200

    fix: decoration for newsletters with useless urls

commit a54bdedcd71995f7ab36d0d0067addac8d98e441
Author: Adam Butler <ab@adam-butler.com>
Date:   Tue Apr 23 17:43:28 2024 +0200

    fix: newsletter layout

commit f482bdda8e237cbcabb8cb467577a37cfeb8813f
Author: Adam Butler <ab@adam-butler.com>
Date:   Fri Apr 19 17:52:19 2024 +0200

    chore: remove the red herring

commit 9b4bcd0350eb235853a3d5394a7ea22ca59569d5
Author: Adam Butler <ab@adam-butler.com>
Date:   Fri Apr 19 17:47:01 2024 +0200

    chore: try to be more resilient to tooManyChanges error

commit 6c150f1632a50f3842a4125c7760b0475ee773af
Author: Adam Butler <ab@adam-butler.com>
Date:   Fri Apr 19 16:30:29 2024 +0200

    all good, but still doesn't work on release

commit f0d1fb573e181f5f6409d7e5bfb7dc21544bc0b4
Author: Adam Butler <ab@adam-butler.com>
Date:   Mon Mar 25 17:57:02 2024 +0100

    more

commit cfdebca4cd55163410e1896974308666ab3fe6a6
Author: Adam Butler <ab@adam-butler.com>
Date:   Thu Mar 21 17:57:42 2024 +0100

    more better

+++ utils/item-utils.js
@@ -2,7 +2,7 @@ const fuzz = require('fuzzball')
 import * as FileSystem from 'expo-file-system'
 const sanitizeHtml = require('sanitize-html')
 import {createItemStyles, compressStyles, expandStyles} from './createItemStyles'
-import {getCachedCoverImagePath} from './index'
+import {fileExists, getCachedCoverImagePath} from './index'
 import log from './log'
 import LZString from 'lz-string'
 import { Platform } from 'react-native'
@@ -34,7 +34,7 @@ export function deflateItem (item) {
     author: item.author,
     coverImageUrl: item.coverImageUrl, // needed by the feed component
     content_length: item.content_length || (item.content_html
-      ? item.content_html.length
+      ? stripTags(item.content_html).length
       : 0),
     created_at: item.created_at,
     date_published: item.date_published,
@@ -47,6 +47,7 @@ export function deflateItem (item) {
     imageDimensions: item.imageDimensions,
     isDecorated: item.isDecorated,
     isExternal: item.isExternal,
+    isNewsletter: item.isNewsletter,
     id: item.id, // needed to match existing copy in store
     readAt: item.readAt,
     // styles: item.styles,
@@ -94,8 +95,9 @@ export function fixRelativePaths (item) {
 
 export function sanitizeContent (item) {
   const settings = {
-    allowedTags: false,
-    allowedAttributes: false
+    allowedTags: sanitizeHtml.defaults.allowedTags.concat([ 'img', 'script', 'video' ]),
+    allowedAttributes: false,
+    allowVulnerableTags: true
   }
   if (item.content_html) item.content_html = sanitizeHtml(item.content_html, settings)
   if (item.content_mercury) item.content_mercury = sanitizeHtml(item.content_mercury, settings)
@@ -116,7 +118,7 @@ export function addMercuryStuffToItem (item, mercury) {
     item = {
       ...item,
       title: mercury.title,
-      content_mercury: mercury.content ? mercury.content : '',
+      content_mercury: mercury.content,
       // body: mercury.content ? mercury.content : '',
       date_published: mercury.date_published,
       author: mercury.author,
@@ -275,14 +277,17 @@ export async function removeCachedCoverImages (items) {
 
   if (Platform.OS === 'web') return
 
-  items.forEach(async (item) => {
+  for (let item of items) {
     let path = getCachedCoverImagePath(item)
     if (path) {
       try {
-        await FileSystem.deleteAsync(path)
+        const exists = await fileExists(path)
+        if (exists) {
+          await FileSystem.deleteAsync(path)
+        }
       } catch (err) {
-        console.log(err)
+        log('removeCachedCoverImages', err)
       }
     }
-  })  
+  }
 }
